requirements.txt
# Core
numpy>=1.21.0
opencv-python>=4.5.5
torch>=1.12.0
torchvision>=0.13.0

# Training
albumentations>=1.3.0
pytorch-lightning>=1.9.0
wandb>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.2
pandas>=1.4.0

# Model Optimization
onnx>=1.12.0
onnxruntime-gpu>=1.12.0
tensorrt>=8.5.0.12
pycuda>=2022.1

# API
fastapi>=0.85.0
uvicorn>=0.19.0
python-multipart>=0.0.5
pydantic>=1.10.0

# Monitoring
psutil>=5.9.0
gpustat>=1.0.0
pynvml>=11.5.0

# Testing
pytest>=7.1.2
pytest-cov>=3.0.0

# Utils
tqdm>=4.64.0
loguru>=0.6.0
pyyaml>=6.0
